fix(access-manager): fix repository delete queries and PermissionService compilation
    - Fix Kotlin type mismatch in PermissionService.kt delete methods.
    - Add @Modifying and explicit JPQL @Query to OwnershipsRepository and SharesRepository to fix deletion ClassCastException.

+++ src/main/kotlin/com/grupo14IngSis/snippetSearcherAccessManager/service/PermissionService.kt
@@ -109,20 +109,20 @@ class PermissionService(
     ): Int {
         val deletedOwnerships = ownershipsRepository.deleteByOwnerIdAndSnippetId(ownerId = userId, snippetId = snippetId)
         val deletedShares = sharesRepository.deleteByUserIdAndSnippetId(snippetId = snippetId, userId = userId)
-        return (deletedOwnerships + deletedShares).toLong()
+        return deletedOwnerships + deletedShares
     }
 
     @Transactional
     fun deletePermissionForUser(userId: String): Int {
         val deletedOwnerships = ownershipsRepository.deleteByOwnerId(ownerId = userId)
         val deletedShares = sharesRepository.deleteByUserId(userId = userId)
-        return (deletedOwnerships + deletedShares).toLong()
+        return deletedOwnerships + deletedShares
     }
 
     @Transactional
     fun deletePermissionForSnippet(snippetId: String): Int {
         val deletedOwnerships = ownershipsRepository.deleteBySnippetId(snippetId = snippetId)
         val deletedShares = sharesRepository.deleteBySnippetId(snippetId = snippetId)
-        return (deletedOwnerships + deletedShares).toLong()
+        return deletedOwnerships + deletedShares
     }
 }
